+++ src/vue/modules/movements-history/index.vue
@@ -68,22 +68,34 @@ export default {
       type: String,
       default: '',
     },
+    /**
+     * Polling config
+     *
+     * {Object} pollConfig - config
+     * {Number} pollConfig.interval - polling interval (ms)
+     * [Function] pollConfig.callback - polling callback
+     */
+    pollConfig: {
+      type: Object,
+      required: false,
+      default: null,
+    },
   },
   data: _ => ({
     isInitialized: false,
     isMovementsLoaded: false,
     isMovementsLoadFailed: false,
+    firstPageLoader: () => {},
   }),
   computed: {
     ...mapGetters('movements-history', {
       balances: types.balances,
       movements: types.movements,
     }),
-    firstPageLoader () {
-      const assetCode = this.assetCode // HACK: passing this.assetCode directly
-      // to function will lead to losing reactivity
-
-      return _ => this.loadMovementsFirstPage(assetCode)
+  },
+  watch: {
+    assetCode (value) {
+      this.initFirstPageLoader()
     },
   },
   async created () {
@@ -91,7 +103,12 @@ export default {
 
     this.setAccountId(this.wallet.accountId)
     await this.loadBalances()
+    this.initFirstPageLoader()
     this.isInitialized = true
+
+    // if (this.pollConfig) {
+    //   this.initPolling()
+    // }
   },
   methods: {
     ...mapMutations('movements-history', {
@@ -103,6 +120,9 @@ export default {
       loadMovements: types.LOAD_MOVEMENTS,
       loadBalances: types.LOAD_BALANCES,
     }),
+    initFirstPageLoader () {
+      this.firstPageLoader = () => this.loadMovementsFirstPage(this.assetCode)
+    },
     async loadMovementsFirstPage (assetCode) {
       this.isMovementsLoaded = false
       try {
@@ -113,6 +133,15 @@ export default {
         ErrorHandler.processWithoutFeedback(e)
         this.isMovementsLoadFailed = true
       }
+    },
+    async initPolling () {
+      setTimeout(async () => {
+        this.firstPageLoader = () => this.loadMovementsFirstPage(this.assetCode)
+        await this.initPolling()
+        if (this.pollConfig.callback) {
+          this.pollConfig.callback(this.movements)
+        }
+      }, this.pollConfig.interval)
     },
   },
 }